refactor(integrations): split manager tests for GH727
